+++ Pod/Classes/ViewControllers/ARBidFlowViewController.m
@@ -4,11 +4,22 @@
 
 - (instancetype)initWithArtworkID:(NSString *)artworkID saleID:(NSString *)saleID
 {
+    return [self initWithArtworkID:artworkID saleID:saleID intent:ARBidFlowViewControllerIntentBid];
+}
+
+- (instancetype)initWithArtworkID:(NSString *)artworkID saleID:(NSString *)saleID intent:(ARBidFlowViewControllerIntent)intent
+{
+  NSDictionary *props = @{
+                          @"artworkID": artworkID ?: [NSNull null],
+                          @"saleID": saleID,
+                          @"intent": (intent == ARBidFlowViewControllerIntentBid ? @"bid" : @"register")
+  };
   if ((self = [super initWithEmission:nil
                            moduleName:@"BidFlow"
-                    initialProperties:@{ @"artworkID": artworkID, @"saleID": saleID }])) {
+                    initialProperties:props])) {
     _artworkID = artworkID;
     _saleID = saleID;
+    _intent = intent;
   }
   return self;
 }