+++ src/lib/AppRegistry.tsx
@@ -89,8 +89,9 @@ const Conversation: React.SFC<ConversationProps> = track<ConversationProps>(prop
 const MyProfile: React.SFC<{}> = () => <MyProfileRenderer render={renderWithLoadProgress(Containers.MyProfile)} />
 
 interface BidFlowProps {
-  artworkID: string
+  artworkID?: string
   saleID: string
+  intent: "bid" | "register"
 }
 const BidFlow: React.SFC<BidFlowProps> = props => (
   <BidFlowRenderer {...props} render={renderWithLoadProgress(Containers.BidFlow)} />