+++ src/lib/Components/Bidding/__stories__/BidFlow.story.tsx
@@ -40,13 +40,21 @@ const BidFlowStoryRenderer: React.SFC<any> = ({ render, query, saleArtworkID }) 
 
 storiesOf("Bidding")
   .add("Show bid flow", () => {
-    return <BidFlowRenderer render={renderWithLoadProgress(BidFlow)} artworkID={testArtworkID} saleID={testSaleID} />
+    return (
+      <BidFlowRenderer
+        render={renderWithLoadProgress(BidFlow)}
+        artworkID={testArtworkID}
+        saleID={testSaleID}
+        intent="bid"
+      />
+    )
   })
   .add("Select Max Bid", () => (
     <BidFlowStoryRenderer
       render={renderWithLoadProgress(MaxBidScreen)}
       query={selectMaxBidQuery}
       saleArtworkID={testSaleArtworkID}
+      intent="bid"
     />
   ))
   .add("Confirm Bid", () => {
@@ -109,6 +117,7 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })
@@ -130,6 +139,7 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })
@@ -154,6 +164,7 @@ storiesOf("Bidding")
         })}
         query={bidResultQuery}
         saleArtworkID={testSaleArtworkID}
+        intent="bid"
       />
     )
   })